Lift codec::common into parent module
